Refactor the project
