+++ src/controllers/index.js
@@ -1,0 +1,7 @@
+const { GCController } = require('./GCController');
+const { LGUController } = require('./LGUController');
+
+module.exports = {
+  GCController,
+  LGUController,
+};
